0214 - 관리자가 아니면 admin 페이지에 접근할 수 없도록 구현

+++ src/admin/admin.controller.ts
@@ -7,7 +7,9 @@ import {
   Post,
   Res,
   Session,
+  UnauthorizedException,
 } from '@nestjs/common';
+import session from 'express-session';
 import { TopicReserve } from 'src/topic/entity/topic-reservation.entity';
 import { Topic } from 'src/topic/entity/topic.entity';
 import { TopicService } from 'src/topic/topic.service';
@@ -27,14 +29,17 @@ export class AdminController {
   @Get('/')
   async main(@Res() res, @Session() session: Record<string, any>) {
     if (!(await this.adminService.checkAdmin(session))) {
-      return res.redirect('/');
+      throw new UnauthorizedException();
     }
 
     return res.render('admin/main');
   }
 
   @Get('/topic')
-  async topic(@Res() res) {
+  async topic(@Res() res, @Session() session: Record<string, any>) {
+    if (!(await this.adminService.checkAdmin(session))) {
+      throw new UnauthorizedException();
+    }
     // test!!!!!!
     //await this.topicService.addTestData();
 
@@ -50,14 +55,20 @@ export class AdminController {
 
   /** 주제 조회 */
   @Get('/topic/:id')
-  async getTopic(@Param('id') id: number, @Res() res) {
+  async getTopic(@Param('id') id: number, @Res() res, @Session() session) {
+    if (!(await this.adminService.checkAdmin(session))) {
+      throw new UnauthorizedException();
+    }
     const topic: Topic = await this.topicService.findOneTopic(id);
     return res.render('admin/topic_detail', { topic: topic });
   }
 
   /** 주제 추가 */
   @Post('new-topic')
-  async createNewTopic(@Body() body: TopicDto, @Res() res) {
+  async createNewTopic(@Body() body: TopicDto, @Res() res, @Session() session) {
+    if (!(await this.adminService.checkAdmin(session))) {
+      throw new UnauthorizedException();
+    }
     this.logger.debug(`body: ${JSON.stringify(body, null, 4)}`);
     await this.topicService.createTopic(body);
     return res.redirect('/admin/topic');
@@ -69,7 +80,11 @@ export class AdminController {
     @Param('id') id: number,
     @Body() body: TopicDto,
     @Res() res,
+    @Session() session,
   ) {
+    if (!(await this.adminService.checkAdmin(session))) {
+      throw new UnauthorizedException();
+    }
     this.logger.debug(`id: ${id}, body: ${JSON.stringify(body, null, 4)}`);
     await this.topicService.updateTopic(id, body);
     res.redirect('/admin/topic');
@@ -77,14 +92,20 @@ export class AdminController {
 
   /** 주제 삭제 */
   @Post('delete-topic/:id')
-  async deleteTopic(@Param('id') id: number, @Res() res) {
+  async deleteTopic(@Param('id') id: number, @Res() res, @Session() session) {
+    if (!(await this.adminService.checkAdmin(session))) {
+      throw new UnauthorizedException();
+    }
     await this.topicService.deleteTopic(id);
     res.redirect('/admin/topic');
   }
 
   /** 예약 조회 */
   @Get('/reserve/:id')
-  async getReserve(@Param('id') id: number, @Res() res) {
+  async getReserve(@Param('id') id: number, @Res() res, @Session() session) {
+    if (!(await this.adminService.checkAdmin(session))) {
+      throw new UnauthorizedException();
+    }
     const reserve: TopicReserve =
       await this.topicService.findOneTopicReserveWithTopic(id);
     this.logger.debug(`reserve: ${JSON.stringify(reserve)}`);
@@ -97,7 +118,14 @@ export class AdminController {
 
   /** 예약 추가 */
   @Post('new-reserve')
-  async createNewReserve(@Body() body: ReserveDto, @Res() res) {
+  async createNewReserve(
+    @Body() body: ReserveDto,
+    @Res() res,
+    @Session() session,
+  ) {
+    if (!(await this.adminService.checkAdmin(session))) {
+      throw new UnauthorizedException();
+    }
     this.logger.debug(`body: ${JSON.stringify(body, null, 4)}`);
     await this.topicService.createReserve(body);
     res.redirect('/admin/topic');
@@ -108,14 +136,21 @@ export class AdminController {
     @Param('id') id: number,
     @Body() body: ReserveDto,
     @Res() res,
+    @Session() session,
   ) {
+    if (!(await this.adminService.checkAdmin(session))) {
+      throw new UnauthorizedException();
+    }
     this.logger.debug(`id: ${id}, body: ${JSON.stringify(body, null, 4)}`);
     await this.topicService.updateReserve(id, body);
     res.redirect('/admin/topic');
   }
 
   @Post('delete-reserve/:id')
-  async deleteRserve(@Param('id') id: number, @Res() res) {
+  async deleteRserve(@Param('id') id: number, @Res() res, @Session() session) {
+    if (!(await this.adminService.checkAdmin(session))) {
+      throw new UnauthorizedException();
+    }
     this.logger.debug(`id: ${id}`);
 
     await this.topicService.deleteReserve(id);
